Changes in version 0.9.5:
Selection of individual tweets and threading with /thread, /reply, /delete and /url, along with @ markers on tweets that are part of a thread.
-noratelimit and -notrack to disable rate limit checks and tracking keywords, respectively, on systems that don't support them (most notably Laconi.ca/Identi.ca).
API addition with $choosecolour.
UTF-8 characters are now allowed in tracking keywords.
Faster and more reliable JSON fetch and parsing method.
Expanded /help text.
Bogus colour warnings when using -noansi are fixed.